linux: Update arch interface
